Only remove directories that exist. This will stop a segmentation fault that occurs if the user exits before davinci makes the temporary directory.

+++ davinci/trunk/system.c
@@ -487,14 +487,17 @@ void rmrf(char *path)
 	char target[1024];
 
 	dir = opendir(path);
-	while((d = readdir(dir)) != NULL) {
-		if (strcmp(d->d_name, ".") && strcmp(d->d_name,"..")) {
-			sprintf(target, "%s/%s", path, d->d_name);
-			unlink(target);
-		}
-	}
-	closedir(dir);
-	rmdir(path);
+   /* eandres: Only delete it if it exists. */
+   if (dir != NULL) {
+	   while((d = readdir(dir)) != NULL) {
+	   	if (strcmp(d->d_name, ".") && strcmp(d->d_name,"..")) {
+	   		sprintf(target, "%s/%s", path, d->d_name);
+	   		unlink(target);
+	   	}
+	   }
+	   closedir(dir);
+	   rmdir(path);
+   }
 }
 
 /*
